NEED TO SUBMIT SO THIS IS IT

+++ EX04_01/EX04_01/Source.cpp
@@ -34,8 +34,8 @@ int main()
 	//cout << endl;
 
 	//compute sum and then average of values in array
-	int sum = 0;
-	int avg = 0;
+	double sum = 0;
+	double avg = 0;
 	for (int i = 0; i < s;i++)//sum of array values
 	{
 		sum += array[i];
